Move assertion parser out
